requirements.txt
discord.py==2.4.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
